UUID + Basic Garbage Collection
Users that access a control point are now assigned a random UUID in a
cookie to ensure that only that user can control the joint at any given
time. Garbage collection works by using datetime and removing any user
who has been active for a given period. The garbage collector and
expiry update functions need to be bound to actions.